fix: moved cmd to higher layer
