+++ cmd/app/generate_all.go
@@ -15,9 +15,9 @@
 package app
 
 import (
+	"github.com/AppsFlyer/terra-crust/cmd/types"
 	"github.com/spf13/cobra"
 
-	"github.com/AppsFlyer/terra-crust/internal/cmd/types"
 	template_reader "github.com/AppsFlyer/terra-crust/internal/services/drivers/template_reader"
 	version_control "github.com/AppsFlyer/terra-crust/internal/services/drivers/version_control"
 )
