+++ cmd/app/generate_main.go
@@ -15,7 +15,7 @@
 package app
 
 import (
-	"github.com/AppsFlyer/terra-crust/internal/cmd/types"
+	"github.com/AppsFlyer/terra-crust/cmd/types"
 	template_reader "github.com/AppsFlyer/terra-crust/internal/services/drivers/template_reader"
 	version_control "github.com/AppsFlyer/terra-crust/internal/services/drivers/version_control"
 	"github.com/spf13/cobra"
